many rewrites and stuff

+++ NetworkTest/Ini/NetworkSettingsIni.cs
@@ -7,6 +7,7 @@ namespace NetworkTest.Ini;
 internal static class NetworkSettingsIni
 {
     static readonly FileIniDataParser parser = new();
+    static bool AlreadyRead;
     static readonly string FileName = "network_settings.ini";
 
     public static void Connect()
@@ -21,16 +22,21 @@ internal static class NetworkSettingsIni
 
     private static NetworkSettings NetworkSettings_OnGet()
     {
+        if (AlreadyRead)
+            return NetworkSettings._instance;
+
         string path = Path.Combine(AppContext.BaseDirectory, FileName);
         if (!File.Exists(path))
         {
             Directory.CreateDirectory(Path.GetDirectoryName(path)!);
             IniData parsedData = IniSourceGeneration.WriteNetworkSettingsIniData(NetworkSettings._instance);
             parser.WriteFile(path, parsedData, System.Text.Encoding.UTF8);
+            AlreadyRead = true;
             return NetworkSettings._instance;
         }
         IniData data = parser.ReadFile(path, System.Text.Encoding.UTF8);
         IniSourceGeneration.ReadNetworkSettingsIniData(NetworkSettings._instance, data);
+        AlreadyRead = true;
         return NetworkSettings._instance;
     }
 }
